Rename em27 module => spectrometer
