+++ finesse/hardware/plugins/spectrometer/__init__.py
