+++ finesse/hardware/plugins/spectrometer/diag_autom.htm
