+++ finesse/hardware/plugins/spectrometer/em27_sensors.py
