+++ finesse/hardware/plugins/spectrometer/opus_interface_base.py
